skills/tinyfy/requirements.txt
tinify>=1.7.1
python-dotenv>=1.0.0
Pillow>=10.0.0
pytest>=8.0.0
